# perf(async): Use tokio_sync's channels instead of futures