--- a/Cargo.toml
+++ b/Cargo.toml
@@ -33,6 +33,7 @@ tokio-executor = "0.1"
 tokio-tcp = "0.1"
 tokio-io = "0.1"
 tokio-codec = "0.1"
+tokio-sync = "0.1"
 tokio-uds = { version = "0.2", optional = true }
 
 [dev-dependencies]